Updated VS solution and module
