--- external/source/exploits/CVE-2018-8440/ALPC-TaskSched-LPE/ClassDiagram.cd
@@ -1,2 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ClassDiagram />
