--- external/source/exploits/CVE-2018-8440/ALPC-TaskSched-LPE/stdafx.cpp
@@ -1,8 +1,0 @@
-// stdafx.cpp : source file that includes just the standard includes
-// $safeprojectname$.pch will be the pre-compiled header
-// stdafx.obj will contain the pre-compiled type information
-
-#include "stdafx.h"
-
-// TODO: reference any additional headers you need in STDAFX.H
-// and not in this file
