+++ external/source/exploits/CVE-2018-8440/dll/ALPC-TaskSched-LPE/stdafx.h
@@ -1,0 +1,15 @@
+// stdafx.h : include file for standard system include files,
+// or project specific include files that are used frequently, but
+// are changed infrequently
+//
+
+#pragma once
+
+
+
+#define RPC_USE_NATIVE_WCHAR
+
+#include <stdio.h>
+#include <tchar.h>
+#include <Windows.h>
+#include <memory>
